REBCNT/REBINT use int_fast32_t
The community build of open-sourced R3-Alpha had a 64-bit conversion
(which has been progressively fixed for correctness in Ren-C).  However
that only fixed it so that 64-bits of address space could be used on
64-bit systems.  It did not convert the indexing of Rebol series to
use 64-bit counts on 64-bit systems.

Preparing the code to handle that would be somewhat involved, as
anywhere APIs are used that take smaller integers--as they relate to
Rebol indexing--could have a problem.  It's not terribly urgent to
do such a conversion.

However--an easier thing to do is to allow Rebol to keep using indexes
that aren't larger than 32-bits, but stored in the fastest variable
type for handling 32-bit numbers on that platform.  For some 64-bit
platforms that means using 64-bit numbers, as having values aligned
on 64-bit boundaries gives a boost.

This goes ahead and does that.  It required any code that was using
REBCNT as a name for a 32-bit unsigned int (or REBINT as a name for
a 32-bit signed int) to switch to uint32_t and int32_t (the types from
stdint.h) in order to keep working.  The cases where this was
a requirement were in things like file formats (BMP, GIF) and the
implementation code for the "deci" datatype (used by MONEY!)

+++ src/include/reb-defs.h
@@ -67,9 +67,13 @@ typedef unsigned char REBYTE; // don't change to uint8_t, see note
 // machines, and a possible loss of performance for forcing a platform to use
 // a specific size int (instead of deferring to C's generic `int`).
 //
-typedef int32_t REBINT; // 32 bit signed integer
-typedef uint32_t REBCNT; // 32 bit (counting number, length in "units")
-typedef uint32_t REBSIZ; // 32 bit (size in bytes)
+// Hence Ren-C switches to using indexes that are provided by <stdint.h> (or
+// the stub "pstdint.h") that are deemed by the compiler to be the fastest
+// representation for 32-bit integers...even if that might be larger.
+//
+typedef int_fast32_t REBINT; // series index, signed, at *least* 32 bits
+typedef uint_fast32_t REBCNT; // series length, unsigned, at *least* 32 bits
+typedef size_t REBSIZ; // 32 bit (size in bytes)
 typedef int64_t REBI64; // 64 bit integer
 typedef uint64_t REBU64; // 64 bit unsigned integer
 typedef float REBD32; // 32 bit decimal
